### 2018-05-02 #713 @vsoch
Thanks @whedon and :wave: @arfon ! The article proof looks good!
