### 2018-05-07 #713 @karthik
@arfon @vsoch I did an initial review and it looks like this would be better suited for someone like @katyhuff or @jakevdp if they have the bandwidth.
