### 2018-05-31 #713 @vsoch
hey @arfon @katyhuff @jakevdp this is almost a month old... ping!
