### 2018-06-11 #713 @arfon
Hi @vsoch - sorry this is taking a while. While we're waiting for an editor to pick this up could you take a look at the reviewer list and suggest a few folks? https://bit.ly/joss-reviewers
