### 2018-06-11 #713 @noamross
I am not available to review, and in any case Python is not one of my programming languages.
