### 2018-06-11 #713 @adamhsparks
Looks nice, but I'm a one trick pony for programming languages and Python isn't one of them.
