### 2018-06-13 #713 @diegoenry
I’m up to reviewing, just let me know the deadlines.

Envoyé de mon iPhone

Le 12 juin 2018 à 21:50, Vanessa Sochat <notifications@github.com<mailto:notifications@github.com>> a écrit :

@arfon<https://nam05.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Farfon&data=02%7C01%7C%7Caf977cc10c0044b46db208d5d09db5ab%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C636644298126072060&sdata=3cmyItaGgejbsYxsdZpAcvstZ5ELxT1wygNK5yqJx2Y%3D&reserved=0> would @iljah<https://nam05.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Filjah&data=02%7C01%7C%7Caf977cc10c0044b46db208d5d09db5ab%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C636644298126072060&sdata=QdbZlGvW%2BwWB%2FSzXysOMaOWQns9jOUFpqUrBIokXEJQ%3D&reserved=0> or @rougier<https://nam05.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Frougier&data=02%7C01%7C%7Caf977cc10c0044b46db208d5d09db5ab%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C636644298126072060&sdata=1TKjuOfgvfLWJNtmR7%2BitQijV0odj3WAtc5A7bBfloI%3D&reserved=0> be suitable to start the review?
