### 2018-06-13 #713 @vsoch
@arfon that's three!! :D @diegoenry I would bet the deadlines are pretty flexible, it's already been a month and we are still at timepoint zero, haha :) I am so grateful for your offering to do the review, and although it's not my decision, I would advocate for giving the time that is needed.
