### 2018-06-13 #775 @whedon
[REVIEW]: HelpMe Command Line Helper Utility
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/vsoch/helpme" target ="_blank">https://www.github.com/vsoch/helpme</a>
**Version:** 0.0.35
