### 2018-06-14 #775 @iljah
Paper.md has statement of need but it's missing from readme.md (Do the authors clearly state what problems the software is designed to solve and who the target audience is?).
