### 2018-06-14 #775 @iljah
Readme.md is missing community guidelines (Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support)
