### 2018-06-14 392ccad
fixing bug with Github issue submission
