+++ docs/_posts/helpers/2018-04-18-github.md
@@ -10,14 +10,13 @@ tags: [helpers]
 ---
 
 Github is the default client. To use Github, 
-<a href="https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line" target="_blank">
-create a personal access token</a>. You can call it whatever you like (`helpme-client` in the example below) 
-and request permissions for "repo". "repo" is the minimum requirement that is needed in order to post issues
-on public repositories. When you click "Generate Token", you will get the token. To use
-it, export it to your environment:
+<a href="https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/" target="_blank">
+create a personal access token</a>. You can call it whatever you like (`helpme-client` in the example below).
+When you generate the application, you will be redirected to a screen that shows the information,
+along with a client id and client secret. You should export both of these to your environment.
 
 ```bash
-export HELPME_GITHUB_TOKEN=xxxxx
+export HELPME_GITHUB_TOKEN=xxxxxxxxxxxxxx
 ```
 
 The token will be found the first time you run the client, and cached in your `$HOME/helpme.cfg`
