### 2018-06-14 cedd765
fixing bug that helpme --help asks for token to close #19
