+++ helpme/version.py
@@ -46,6 +46,7 @@ LICENSE = "LICENSE"
 
 INSTALL_REQUIRES = (
     ('requests', {'min_version': '2.18.4'}),
+    ('asciinema', {'min_version': '2.0.1'}),
 )
 
 ################################################################################
