### 2018-06-14 3bb2ec5
adding statement of need to README.md
