### 2018-06-14 7686336
adding CONTRIBUTING guidelines to README
