### 2018-06-14 #775 @vsoch
hey @iljah ! I just merged some changes to address your points above:
 - I added a "year" field to references, where appropriate
 - The version of the software coincides with the pypi release. I can definitely create a tag/official release if you think this is necessary, but generally I rely on pypi to provide these files (so there isn't confusion/ doubling of it)
 - I added the statement of need to the README.md.
 - I added community guidelines to the README.md
