### 2018-06-14 #775 @rougier
I think you need to work on the `paper.bib` since some references does not make much sense. For example, the github reference is not really a reference. It would be better to just add a link in the paper rather than having it as a reference.

Also, you can reformat some references, e.g.:

```
@MISC{helpme,
  title        = "{HelpMe}",
  booktitle    = "{HelpMe}",
  author       = "{vsoch}",
  abstract     = "HelpMe - a command line tool for helping you out.",
  publisher    = "vsoch",
  year        = {2018},
  howpublished = "\url{https://vsoch.github.io/helpme/helpme/}",
  note         = "Accessed: 2018-4-30"
}

could be rewritten as:

@MISC{helpme,
  title =  "{HelpMe}",
  author = "Sochat, Vanessa",
  year =  {2018},
  url = "https://vsoch.github.io/helpme/helpme/",
}
```
