### 2018-06-14 #775 @rougier
For the README, why don't you advertise the pip install method instead of the docker one ?
