### 2018-06-14 59a5157
tweaking readme, and improving paper bib
