+++ paper/paper.bib
@@ -1,33 +1,27 @@
 @MISC{helpme,
-  title        = "{HelpMe}",
-  booktitle    = "{HelpMe}",
-  author       = "{vsoch}",
-  abstract     = "HelpMe - a command line tool for helping you out.",
-  publisher    = "vsoch",
-  year        = {2018},
-  howpublished = "\url{https://vsoch.github.io/helpme/helpme/}",
-  note         = "Accessed: 2018-4-30"
+  title  =  "{HelpMe Documentation}",
+  author = "Sochat, Vanessa",
+  year   =  {2018},
+  url    = "\url{https://vsoch.github.io/helpme/helpme/}",
 }
 
-
 @MISC{asciinema,
-  title        = "asciinema - Record and share your terminal sessions, the
-                  right way",
-  year        = {2018},
-  howpublished = "\url{https://asciinema.org/}",
-  note         = "Accessed: 2018-4-30"
+  title  = "asciinema - Record and share your terminal sessions, the right way",
+  year   = {2018},
+  author = "Kulik, Marcin"
+  url    = "\url{https://asciinema.org/}",
 }
 
 @MISC{uservoice,
-  title        = "Introducing {UserVoice} Helpdesk, a Support Tool that Doesn't
-                  Suck | {UserVoice} Blog",
-  howpublished = "\url{https://community.uservoice.com/blog/introducing-uservoice-helpdesk-a-support-tool-that-doesnt-suck/}",
-  year        = {2011},
-  note         = "Accessed: 2018-4-30"
+  title  = "Introducing {UserVoice} Helpdesk, a Support Tool that Doesn't Suck",
+  url    = "\url{https://community.uservoice.com/blog/introducing-uservoice-helpdesk-a-support-tool-that-doesnt-suck/}",
+  year   = {2011},
+  author = "Hamilton, Evan"
 }
 
 @MISC{github,
   title       = "Build software better, together",
+  url         = "\url{https://www.github.com/}",
   abstract    = "GitHub is where people build software. More than 27 million
                  people use GitHub to discover, fork, and contribute to over 80
                  million projects.",
@@ -35,11 +29,11 @@
 }
 
 @MISC{software,
-  title       = "helpme",
+  title       = "HelpMe on Github",
+  url         = "\url{https://www.github.com/vsoch/helpme/}",
   author      = "Sochat, Vanessa",
   year        = {2018},
   abstract    = "GitHub is where people build software. More than 27 million
                  people use GitHub to discover, fork, and contribute to over 80
                  million projects.",
-  institution = "Github"
 }
