### 2018-06-14 #775 @vsoch
Okay, I made some changes to the bib, please take a look!

I also added the install with pip to the readme. It wasn't there mostly because I want users to rely on the documentation base, but I agree that some won't want to go that many clicks (and the README is a good place to have it).
