+++ helpme/main/github/__init__.py
@@ -65,14 +65,14 @@ class Helper(HelperBase):
            
            {'user_prompt_issue': 'I want to do the thing.', 
             'user_prompt_repo': 'vsoch/hello-world', 
+            'user_prompt_title': 'Error with this thing',
             'record_asciinema': '/tmp/helpme.93o__nt5.json',
             'record_environment': ((1,1),(2,2)...(N,N))}
 
            self.token should be propogated with the personal access token
-        '''
- 
-        title = "HelpMe Github Issue: %s" %(self.run_id)
+        ''' 
         body = self.data['user_prompt_issue']
+        title = self.data['user_prompt_title']
         repo = self.data['user_prompt_repo']
 
         # Step 1: Environment
@@ -96,7 +96,8 @@ class Helper(HelperBase):
  
         # Add other metadata about client
 
-        body += "\ngenerated by [HelpMe](https://vsoch.github.io/helpme/)"
+        body += "\n\ngenerated by [HelpMe](https://vsoch.github.io/helpme/)"
+        body += "\nHelpMe Github Issue: %s" %(self.run_id)
 
         # Submit the issue
 
