+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/sregistry-cli/tree/master) (0.0.x)
+ - adding option for user to specify title for issue (0.0.32) 
  - import of Helper not necessary (then asked for Github token) (0.0.31) 
  - instructions must be for Github app (personal token not scoped correctly) (0.0.3) 
  - first alpha release with uservoice and github helpers (0.0.2)
