+++ helpme/helpme.cfg
@@ -4,6 +4,7 @@
 user_message_welcome = Hello Friend! I'll help you post a Github issue today.
 user_prompt_repo = What is the Github repository to post to? (e.g., username/reponame)
 user_prompt_issue = Please describe the issue you are having 
+user_prompt_title = Please provide a meaningful title for your issue
 record_asciinema = True
 record_environment = True
 
