+++ paper/paper.bib
@@ -8,7 +8,7 @@
 @MISC{asciinema,
   title  = "asciinema - Record and share your terminal sessions, the right way",
   year   = {2018},
-  author = "Kulik, Marcin"
+  author = "Kulik, Marcin",
   url    = "\url{https://asciinema.org/}",
 }
 
