+++ paper/paper.bib
@@ -2,26 +2,26 @@
   title  =  "{HelpMe Documentation}",
   author = "Sochat, Vanessa",
   year   =  {2018},
-  url    = "\url{https://vsoch.github.io/helpme/helpme/}",
+  url    = "https://vsoch.github.io/helpme/helpme/",
 }
 
 @MISC{asciinema,
   title  = "asciinema - Record and share your terminal sessions, the right way",
   year   = {2018},
   author = "Kulik, Marcin",
-  url    = "\url{https://asciinema.org/}",
+  url    = "https://asciinema.org/",
 }
 
 @MISC{uservoice,
   title  = "Introducing {UserVoice} Helpdesk, a Support Tool that Doesn't Suck",
-  url    = "\url{https://community.uservoice.com/blog/introducing-uservoice-helpdesk-a-support-tool-that-doesnt-suck/}",
+  url    = "https://community.uservoice.com/blog/introducing-uservoice-helpdesk-a-support-tool-that-doesnt-suck/",
   year   = {2011},
   author = "Hamilton, Evan"
 }
 
 @MISC{github,
   title       = "Build software better, together",
-  url         = "\url{https://www.github.com/}",
+  url         = "https://www.github.com/",
   abstract    = "GitHub is where people build software. More than 27 million
                  people use GitHub to discover, fork, and contribute to over 80
                  million projects.",
@@ -30,7 +30,7 @@
 
 @MISC{software,
   title       = "HelpMe on Github",
-  url         = "\url{https://www.github.com/vsoch/helpme/}",
+  url         = "https://www.github.com/vsoch/helpme/",
   author      = "Sochat, Vanessa",
   year        = {2018},
   abstract    = "GitHub is where people build software. More than 27 million
