### 2018-06-15 #775 @iljah
Pdf looks pretty good except github reference still has n.d., maybe adding a year would make that go away?
