+++ paper/paper.bib
@@ -21,6 +21,7 @@
 
 @MISC{github,
   title       = "Build software better, together",
+  year        = {2015},
   url         = "https://www.github.com/",
   abstract    = "GitHub is where people build software. More than 27 million
                  people use GitHub to discover, fork, and contribute to over 80
