### 2018-06-15 2594ffb
adding manifest
