+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include README.md LICENSE
+recursive-include helpme *
+global-exclude *.py[co]
+global-exclude __pycache__
