+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/sregistry-cli/tree/master) (0.0.x)
+ - fixing package data (0.0.35) 
  - adding MANIFEST.in to include data files (0.0.34) 
  - fixing import bug (0.0.33) 
  - adding option for user to specify title for issue (0.0.32) 
