### 2018-06-16 #775 @rougier
@vsoch Don't you need to make a GitHub release?
