### 2018-06-16 #775 @rougier
For the README, I think all the information is present but the structure is a bit weird. You have the overview and usage near the end, the contributing and license are at the subsection level (instead of section) and it is not clear that local and docker refers to the installation. I guess you can leave it as it is if your prefer but a restructuration might be an option worth to consider.
