### 2018-06-16 #775 @vsoch
Oh thanks @rougier I didn't remember that I needed to do this. I'll fix up the README, create a release, and post here when that is done.
