### 2018-06-20 #775 @vsoch
All set! Here is the DOI: https://zenodo.org/account/settings/github/repository/vsoch/helpme#

I had created the tagged release 2 days ago, but in the wrong order so it didn't trigger anything on Zenodo (this I had completely forgotten about!) I re-created the release and now the two are synced up. Thanks for the heads up @arfon !
